refactor: move clients folder to src
